requirements.txt
pillow>=8.4.0
numpy>=1.19.5